update digital analog results
