+++ src/qilisdk/digital/digital_result.py
@@ -11,7 +11,6 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from abc import abstractmethod
 from pprint import pformat
 
 from qilisdk.common import Result
@@ -21,12 +20,10 @@ from qilisdk.yaml import yaml
 @yaml.register_class
 class DigitalResult(Result):
     """
-    Abstract class representing the result of a quantum circuit measurement.
+    Class representing the result of a quantum circuit measurement.
 
     DigitalResult encapsulates the outcome of a digital measurement performed on a quantum circuit.
-    It includes the total number of measurement shots and mandates that subclasses implement
-    properties and methods to expose the number of qubits involved, the raw measurement samples,
-    and methods to compute the probabilities of measurement outcomes.
+    It includes the total number of measurement shots, the measurement samples and measurement probabilities.
 
     Attributes:
         nshots (int): The number of measurement shots performed in the experiment.
@@ -34,14 +31,36 @@ class DigitalResult(Result):
         samples (dict[str, int]): A dictionary where keys are bitstrings representing measurement outcomes and values are the number of times each outcome was observed.
     """
 
-    def __init__(self, nshots: int) -> None:
+    def __init__(self, nshots: int, samples: dict[str, int]) -> None:
         """
         Initializes a DigitalResult instance.
 
         Args:
             nshots (int): The total number of measurement shots performed.
+            samples (dict[str, int]): A dictionary mapping bitstring outcomes to their occurrence counts.
+                All keys (bitstrings) must have the same length, which determines the number of qubits.
+
+        Raises:
+            ValueError: If the samples dictionary is empty or if not all bitstring keys have the same length.
         """
         self._nshots = nshots
+        self._samples = samples
+
+        # Ensure samples is not empty and is correct.
+        if not samples:
+            raise ValueError("The samples dictionary is empty.")
+        bitstrings = list(samples.keys())
+        nqubits = len(bitstrings[0])
+        if not all(len(bitstring) == nqubits for bitstring in bitstrings):
+            raise ValueError("Not all bitstring keys have the same length.")
+
+        # Assign nqubits to attribute
+        self._nqubits = nqubits
+
+        # Calculate probabilities
+        self._probabilities = {
+            bitstring: counts / self._nshots if self._nshots > 0 else 0.0 for bitstring, counts in self._samples.items()
+        }
 
     @property
     def nshots(self) -> int:
@@ -54,7 +73,6 @@ class DigitalResult(Result):
         return self._nshots
 
     @property
-    @abstractmethod
     def nqubits(self) -> int:
         """
         Gets the number of qubits involved in the measurement.
@@ -62,9 +80,9 @@ class DigitalResult(Result):
         Returns:
             int: The number of qubits measured.
         """
+        return self._nqubits
 
     @property
-    @abstractmethod
     def samples(self) -> dict[str, int]:
         """
         Gets the raw measurement samples.
@@ -73,17 +91,18 @@ class DigitalResult(Result):
             dict[str, int]: A dictionary where keys are bitstrings representing measurement outcomes
             and values are the number of times each outcome was observed.
         """
+        return dict(self._samples)
 
-    @abstractmethod
-    def get_probabilities(self) -> dict[str, float]:
+    @property
+    def probabilities(self) -> dict[str, float]:
         """
-        Computes the probabilities for each measurement outcome.
+        Gets the probabilities for each measurement outcome.
 
         Returns:
             dict[str, float]: A dictionary mapping each bitstring outcome to its corresponding probability.
         """
+        return dict(self._probabilities)
 
-    @abstractmethod
     def get_probability(self, bitstring: str) -> float:
         """
         Computes the probability of a specific measurement outcome.
@@ -94,6 +113,7 @@ class DigitalResult(Result):
         Returns:
             float: The probability of the specified bitstring occurring.
         """
+        return self._probabilities.get(bitstring, 0.0)
 
     def __repr__(self) -> str:
         """
