+++ src/qilisdk/extras/cuda/cuda_analog_result.py
@@ -11,10 +11,9 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
-from qilisdk.digital.digital_result import DigitalResult
+from qilisdk.analog.analog_result import AnalogResult
 from qilisdk.yaml import yaml
 
 
 @yaml.register_class
-class QaaSDigitalResult(DigitalResult): ...
+class CudaAnalogResult(AnalogResult): ...
